bd:cambio

+++ database/migrations/2023_03_28_002953_create_contracts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->nullable()->constrained()->onDelete('set null');
             $table->date('fecha_ini');
             $table->date('fecha_fin');
-            $table->integer('salario');
+            $table->double('salario');
             $table->timestamps();
         });
     }
